trace-cmd: 2.9-dev -> 2.9.1, kernelshark: 1.1.0 -> 1.2
Besides updating the version, this commit decouples trace-cmd and
kernelshark source revisions. These two programs are developed in the
same repository but are released independently. Nixpkgs built the
previous versions of these programs from the same source commit, which
meant that at least one of the programs were not built from the
officially released version. Now we use the release tags of the
corresponding programs.