defaultGemConfig.pg_query: 2.0.1 -> 2.0.2
pg_query 2.0.2 apparently has working ARM support.

This change also re-locks sqlint to use the newer version, and
generalises the platform spec to avoid being unnecessarily
restrictive.

As before, other packages which refer to pg_query are unaffected,
because they pin an older version of pg_query which is handled
separately in the gem-config expressions.

Follow-up to #116785