python3Packages.panel: 0.9.7 -> 0.11.1
Since 0.10.0 setup.py the node assets are no longer included in the
tarball. During setup, npm is used to first collect assets. Then,
a function is bokeh used to bundle the assets. Unfortunately,
at that time it attempts to collect files *again*. This we cannot
handle so we do not bundle.